requirements.txt
pyyaml>=6.0
httpx>=0.28.1
